+++ src/pixelator/cli/main.py
@@ -5,7 +5,9 @@ Copyright © 2022 Pixelgen Technologies AB.
 
 import atexit
 import multiprocessing
+import os
 import sys
+from pathlib import Path
 
 import click
 import yappi
@@ -93,6 +95,16 @@ def main_cli(ctx, verbose: bool, profile: bool, log_file: str, cores: int):
     ctx.obj["LOGGER"] = ctx.with_resource(LoggingSetup(log_file, verbose=verbose))
     ctx.obj["VERBOSE"] = verbose
     ctx.obj["CORES"] = max(1, cores)
+
+    # Read in environment variables
+    duckdb_temp_dir = os.environ.get("PIXELATOR_DUCKDB_TEMP_DIR")
+    if duckdb_temp_dir:
+        ctx.obj["DUCKDB_TEMP_DIR"] = Path(duckdb_temp_dir)
+
+    duckdb_tmp_dir_size = os.environ.get("PIXELATOR_DUCKDB_MAX_TEMP_DIR_SIZE")
+    if duckdb_tmp_dir_size:
+        ctx.obj["DUCKDB_MAX_TEMP_DIR_SIZE"] = duckdb_tmp_dir_size.strip()
+
     return 0