+++ src/pixelator/pna/config/panel.py
@@ -5,7 +5,10 @@ Copyright © 2022 Pixelgen Technologies AB.
 
 from __future__ import annotations
 
+import warnings
+from functools import cached_property
 from pathlib import Path
+from typing import TYPE_CHECKING, List, Optional
 
 try:
     from typing import Self
@@ -13,27 +16,28 @@ except ImportError:
     from typing_extensions import Self
 
 
+import re
 from typing import TYPE_CHECKING, Optional
 
 import pandas as pd
+import ruamel.yaml as yaml
 
-from pixelator.common.config import AntibodyPanel
+from pixelator import read_pna
+from pixelator.common.config import AntibodyPanelMetadata
 from pixelator.common.types import PathType
 from pixelator.common.utils import logger
 
 if TYPE_CHECKING:
-    from pixelator.config.config_class import AntibodyPanelMetadata
     from pixelator.pna.config.config_class import PNAConfig
 
 
-class PNAAntibodyPanel(AntibodyPanel):
+class PNAAntibodyPanel:
     """Class representing a Molecular Pixelation antibody panel."""
 
     # required columns
     _REQUIRED_COLUMNS = [
         "marker_id",
         "control",
-        "nuclear",
         "sequence_1",
         "sequence_2",
         "conj_id",
@@ -97,6 +101,82 @@ class PNAAntibodyPanel(AntibodyPanel):
         return cls(df, metadata, file_name=panel_file.name)
 
     @classmethod
+    def from_pxl(cls, filename: PathType) -> Self:
+        """Create an AntibodyPanel from a pxl file.
+
+        :param filename: The path to the panel file.
+        :returns: The AntibodyPanel object.
+        :raises AssertionError: exception if panel file is missing,
+        :rtype: AntibodyPanel
+        """
+        pxl_file = Path(filename)
+
+        if not pxl_file.is_file() or pxl_file.suffix != ".pxl":
+            raise AssertionError(
+                f"Panel file {filename} not found or has an incorrect format"
+            )
+
+        logger.debug("Creating Antibody panel from file %s", filename)
+
+        pxl_data = read_pna(pxl_file)
+        adata = pxl_data.adata()
+        panel_metadata = adata.uns["panel_metadata"]
+        panel_columns = panel_metadata.pop("panel_columns")
+        df = adata.var.reset_index(names="marker_id")[panel_columns]
+        metadata = AntibodyPanelMetadata.model_validate(panel_metadata)
+
+        logger.debug("Antibody panel from file %s created", filename)
+
+        return cls(df, metadata, file_name=pxl_file.name)
+
+    @property
+    def name(self) -> Optional[str]:
+        """The name defined in the panel metadata."""
+        return self._metadata.name
+
+    @property
+    def version(self) -> Optional[str]:
+        """Return the panel file version."""
+        return self._metadata.version
+
+    @property
+    def description(self) -> Optional[str]:
+        """Return the panel file description."""
+        return self._metadata.description
+
+    @property
+    def aliases(self) -> list[str]:
+        """Return the (optional) list of panel file aliases."""
+        return self._metadata.aliases
+
+    @classmethod
+    def _parse_header(cls, file: Path) -> AntibodyPanelMetadata:
+        """Parse front-matter YAML from a file.
+
+        :param file: the file to parse
+        :return AntibodyPanelMetadata: a pydantic model with the metadata
+        """
+        yaml_loader = yaml.YAML(typ="safe")
+
+        metadata_lines = []
+        with open(str(file), "r") as f:
+            for line in f:
+                if line.startswith("# "):
+                    metadata_lines.append(line[2:])
+                else:
+                    break
+
+        metadata = "".join(metadata_lines)
+        raw_config = list(yaml_loader.load_all(metadata))
+
+        if len(raw_config) == 0:
+            warnings.warn(f"Expected a YAML frontmatter in {file}")
+            return AntibodyPanelMetadata(version=None, name=None, description=None)
+
+        frontmatter = raw_config[0]
+        return AntibodyPanelMetadata.model_validate(frontmatter)
+
+    @classmethod
     def _parse_panel(cls, panel_file: Path) -> pd.DataFrame:
         panel = pd.read_csv(str(panel_file), comment="#")
 
@@ -114,6 +194,51 @@ class PNAAntibodyPanel(AntibodyPanel):
         # return a local copy
         return panel.copy()
 
+    @cached_property
+    def markers_control(self) -> List[str]:
+        """Return a list of marker control (names)."""
+        return list(self._df[self._df["control"]].marker_id.unique())
+
+    @cached_property
+    def markers(self) -> List[str]:
+        """Return the list of unique markers in the panel."""
+        return list(self._df.marker_id.unique())
+
+    @property
+    def df(self) -> pd.DataFrame:
+        """Return the panel dataframe."""
+        return self._df
+
+    @property
+    def filename(self) -> Optional[str]:
+        """Return the filename of the marker panel."""
+        return self._filename
+
+    @cached_property
+    def size(self) -> int:
+        """Return the size of the marker panel."""
+        return self._df.shape[0]
+
+    def get_marker_id(self, seq: str) -> str:
+        """Return the marker name."""
+        return self._df.loc[seq].marker_id
+
+    @staticmethod
+    def _validate_sequences(panel_df, sequence_col):
+        errors = []
+        sequences = panel_df[sequence_col]
+        ref_length = len(sequences.iloc[0])
+        if not sequences.apply(lambda x: len(x) == ref_length).all():
+            errors.append(f"All {sequence_col} values must have the same length.")
+
+        if not sequences.str.match("^[ATCG]*$").all():
+            errors.append(
+                f"All {sequence_col} values must only contain ATCG characters. Offending values: "
+                f"{sequences[~sequences.str.match('^[ATCG]*$')].tolist()}"
+            )
+
+        return errors
+
     @classmethod
     def validate_antibody_panel(cls, panel_df: pd.DataFrame) -> list[str]:
         """Validate the antibody panel dataframe.
@@ -121,7 +246,22 @@ class PNAAntibodyPanel(AntibodyPanel):
         :param panel_df: The dataframe containing the panel information.
         :returns: A list of errors found in the panel.
         """
-        errors = super().validate_antibody_panel(panel_df)
+        errors = []
+
+        # some basic sanity check on the panel size and columns
+        if not set(cls._REQUIRED_COLUMNS).issubset(set(panel_df.columns)):
+            missing_columns = set(cls._REQUIRED_COLUMNS) - set(panel_df.columns)
+            errors.append(f"Panel has missing required columns: {missing_columns}")
+            return errors
+
+        if panel_df.shape[0] == 0:
+            errors.append("Panel file is empty")
+            return errors
+
+        # sanity check on the unique columns
+        for col in cls._UNIQUE_COLUMNS:
+            if not len(panel_df[col].unique()) == len(panel_df[col]):
+                errors.append(f"All values in column: {col} were not unique")
 
         if any(panel_df["marker_id"].str.contains("_")):
             # Markers should not contain underscores since this messes
@@ -131,6 +271,30 @@ class PNAAntibodyPanel(AntibodyPanel):
                 "Please use dashes instead. Offending values: "
                 f"{panel_df['marker_id'][panel_df['marker_id'].str.contains('_')]}"
             )
+
+        # Check UniProt IDs format conforming to the UniProt naming convention. Empty IDs are allowed.
+        if "uniprot_id" in panel_df.columns:
+            # Pattern for valid UniProt IDs
+            pattern = r"^[OPQ][0-9][A-Z0-9]{3}[0-9]|[A-NR-Z][0-9]([A-Z][A-Z0-9]{2}[0-9]){1,2}$"
+
+            def check_id(id_str):
+                if pd.isna(id_str):
+                    return True
+                return all(
+                    bool(re.match(pattern, id_)) for id_ in str(id_str).split(";")
+                )
+
+            bad_ids = panel_df[~panel_df["uniprot_id"].apply(check_id)]["uniprot_id"]
+
+            if len(bad_ids) > 0:
+                errors.append(
+                    "Invalid UniProt IDs found."
+                    "Please conform to the naming convention or remove the following IDs:"
+                    f"{bad_ids.tolist()}"
+                )
+
+        errors += cls._validate_sequences(panel_df, "sequence_1")
+        errors += cls._validate_sequences(panel_df, "sequence_2")
 
         return errors