+++ tests/pna/collapse/test_combine_collapse.py
@@ -1,6 +1,7 @@
 """Copyright © 2025 Pixelgen Technologies AB."""
 
 import json
+from pathlib import Path
 
 import pandas as pd
 import pytest
@@ -543,7 +544,12 @@ def marker2_parquet():
 def test_combine_independent_parquet(tmp_path, m1_collapsed_data, m2_collapsed_data):
     combined_parquet = tmp_path / "combined.parquet"
     stats = combine_independent_parquet_files(
-        m1_collapsed_data, m2_collapsed_data, combined_parquet
+        m1_collapsed_data,
+        m2_collapsed_data,
+        combined_parquet,
+        temp_directory=Path("/tmp"),
+        threads=2,
+        memory_limit=2**30,
     )
 
     assert combined_parquet.exists()