+++ tests/pna/config/test_config.py
@@ -6,6 +6,7 @@ Copyright © 2023 Pixelgen Technologies AB.
 
 import copy
 
+import pandas as pd
 import pytest
 
 from pixelator.common.config import (
@@ -18,7 +19,7 @@ from pixelator.pna.config.config_class import (
     load_panels_package,
 )
 from pixelator.pna.config.config_instance import pna_config
-from pixelator.pna.config.panel import load_antibody_panel
+from pixelator.pna.config.panel import PNAAntibodyPanel, load_antibody_panel
 
 
 def test_config_creation():
@@ -104,11 +105,13 @@ def test_loading_panel_from_config(config_with_multiple_versions):
     assert panel.version == "0.2.0"
 
 
-@pytest.mark.parametrize("panel_name", ["proxiome-immuno-155", "proxiome-immuno-155"])
+@pytest.mark.parametrize(
+    "panel_name", ["proxiome-immuno-155-v1", "proxiome-immuno-155-v2"]
+)
 def test_loading_panel_from_config_alias(panel_name):
     panel = pna_config.get_panel(panel_name)
     assert panel.name == panel_name
-    assert panel.version == "1.0.0"
+    assert panel.version == f"{panel_name[-1]}.0.0"
 
 
 def test_loading_panel_from_config_specific_version(config_with_multiple_versions):
@@ -123,7 +126,7 @@ def test_loading_panel_from_config_specific_version(config_with_multiple_version
 
 def test_load_antibody_panel_util(pna_data_root):
     cgf_panel = load_antibody_panel(pna_config, "proxiome-immuno-155")
-    assert cgf_panel.name == "proxiome-immuno-155"
+    assert cgf_panel.name == "proxiome-immuno-155-v2"
 
     path_panel = load_antibody_panel(
         pna_config, pna_data_root / "test-pna-panel-v2.csv"
@@ -135,17 +138,46 @@ def test_load_antibody_panel_util(pna_data_root):
         load_antibody_panel(pna_config, "human-qwdqwdqwdqdw-proteomics")
 
 
+def test_panel_with_non_dna_sequences(pna_data_root):
+    panel_df = pd.read_csv(pna_data_root / "test-pna-panel-v2.csv", skiprows=9)
+    panel_df.loc[0, "sequence_1"] = "PPPPPP"
+    errors = PNAAntibodyPanel.validate_antibody_panel(panel_df)
+    assert len(errors) == 2
+    assert errors[0] == "All sequence_1 values must have the same length."
+    assert (
+        errors[1]
+        == "All sequence_1 values must only contain ATCG characters. Offending values: ['PPPPPP']"
+    )
+
+
 def test_list_panel_names(pna_data_root):
-    assert sorted(pna_config.list_panel_names(include_aliases=True)) == [
-        "proxiome-immuno-155",
-        "proxiome-immuno-155plex",
-        "proxiome-immuno-156-FMC63",
-        "proxiome-immuno-156-FMC63plex",
-    ]
+    assert sorted(pna_config.list_panel_names(include_aliases=True)) == sorted(
+        [
+            "proxiome-immuno-155",
+            "proxiome-immuno-155plex",
+            "proxiome-immuno-155-v1",
+            "proxiome-immuno-155plex-v1",
+            "proxiome-immuno-155-v2",
+            "proxiome-immuno-155plex-v2",
+            "proxiome-immuno-156-FLAG",
+            "proxiome-immuno-156-FLAG-v2",
+            "proxiome-immuno-156-FLAGplex",
+            "proxiome-immuno-156-FLAGplex-v2",
+            "proxiome-immuno-156-FMC63",
+            "proxiome-immuno-156-FMC63plex",
+            "proxiome-immuno-156-FMC63-v1",
+            "proxiome-immuno-156-FMC63-v2",
+            "proxiome-immuno-156-FMC63plex-v1",
+            "proxiome-immuno-156-FMC63plex-v2",
+        ]
+    )
 
     assert sorted(pna_config.list_panel_names(include_aliases=False)) == [
-        "proxiome-immuno-155",
-        "proxiome-immuno-156-FMC63",
+        "proxiome-immuno-155-v1",
+        "proxiome-immuno-155-v2",
+        "proxiome-immuno-156-FLAG-v2",
+        "proxiome-immuno-156-FMC63-v1",
+        "proxiome-immuno-156-FMC63-v2",
     ]